// rss-css-selectors/src/components/types/enums.ts
export enum GLASS {
  TUMBLER = 'tumbler',
  COCKTAIL = 'cocktail',
  BALLOON = 'balloon',
}

export enum COLOR {
  BLUE = 'blue',
  GREEN = 'green',
  PURPLE = 'purple',
  NONE = 'none',
}

export enum BEVERAGE {
  JUICE = 'juice',
  WINE = 'wine',
  MILK = 'milk',
  COLA = 'cola',
  NONE = 'none',
}

export enum FULLNESS {
  FULL = 'full',
  HALF = 'half',
  NONE = 'none',
}